% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{annotation-class}
\alias{annotation}
\alias{annotation-class}
\title{annotation class}
\description{
This class holds an annotation object that defines how annotations relate to
features, as well as various pieces about each annotation
}
\details{
These objects may be created by hand, or may result from specific functions
to create them. Most notably, this package provides functions for creating
a Gene Ontology annotation.
}
\section{Slots}{

\describe{
\item{\code{annotation_features}}{list of annotation to feature relationships}

\item{\code{description}}{character vector providing descriptive text about the annotation}

\item{\code{counts}}{numeric vector of how many features are in each annotation}

\item{\code{links}}{character vector defining html links for each annotation (may be empty)}

\item{\code{type}}{a one word short description of the "type" of annotation}
}}

