% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graph_visualization.R
\docType{methods}
\name{annotation_combinations,cc_graph-method}
\alias{annotation_combinations,cc_graph-method}
\title{unique combinations}
\usage{
\S4method{annotation_combinations}{cc_graph}(in_graph)
}
\arguments{
\item{in_graph}{the \S4Class{cc_graph} to work on}
}
\value{
node_assignment
}
\description{
determine the unique combinations of annotations that exist in the
significant matrix of the \S4Class{cc_graph} and assign each node in the graph
to a group.
}

