% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{cc_graph-class}
\alias{cc_graph}
\alias{cc_graph-class}
\title{cc_graph}
\description{
A \code{cc_graph} class is a \code{graphNEL} with the added slot of
\code{significant}, a matrix of rows (nodes / annotations) and whether
they were found to be significant in a given enrichment (columns). This
matrix is used for classifying the annotations into different groups, and
generating either pie-charts or coloring the nodes in a visualization.
}
\section{Slots}{

\describe{
\item{\code{significant}}{numeric matrix of ones and zeros}
}}

