\name{ccCompareGeneric-methods}
\docType{methods}
\alias{ccCompareGeneric}
\alias{ccCompareGeneric-methods}
\alias{GENccEnrichResult,ccCompareGeneric-method}
\alias{ccOptions,ccCompareGeneric-method}
\title{  Methods for Function \code{ccCompareGeneric} in Package \pkg{categoryCompare} }
\description{
  Methods for function \code{ccCompareGeneric} in package \pkg{categoryCompare} 
}
\section{Methods}{
\describe{

\item{\code{signature(gccResult = "GENccEnrichResult", ccOptions = "ccOptions")}}{
%%  describe this method here
}
}}
\keyword{methods}
\keyword{  other possible keyword(s)  }
