% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combine_enrichments.R
\name{combine_annotation_features}
\alias{combine_annotation_features}
\title{combine annotation-features}
\usage{
combine_annotation_features(annotation_features)
}
\arguments{
\item{annotation_features}{list of annotation_features to combine}
}
\value{
list of combined annotations
}
\description{
For the generation of a proper annotation-annotation relationship graph, we
need to combine the annotation-feature relationships across multiple
\linkS4class{annotation} objects
}

