% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combine_enrichments.R
\docType{methods}
\name{combine_annotations,list-method}
\alias{combine_annotations,list-method}
\title{combine annotations}
\usage{
\S4method{combine_annotations}{list}(annotation_list)
}
\arguments{
\item{...}{one or more \linkS4class{annotation}}
}
\value{
\linkS4class{annotation}
}
\description{
Takes multiple \linkS4class{annotation} objects and combines them so that there
is a consistent sole set for creating the \code{annotation_graph} and providing
other information about each annotation entry.
}

