% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combine_enrichments.R
\docType{methods}
\name{combine_enrichments,enriched_result-method}
\alias{combine_enrichments,enriched_result-method}
\title{combine enrichments}
\usage{
\S4method{combine_enrichments}{enriched_result}(...)
}
\arguments{
\item{...}{one or more \linkS4class{enriched_results}}
}
\value{
\linkS4class{combined_enrichment}
}
\description{
This is one of the primary workhorse functions behind \pkg{categoryCompare}.
The primary function of \code{categoryCompare} is to enable \emph{comparisons}
of different enrichment analyses. To facilitate that, we must first
\strong{combine} one (really, we can do this with a single) or more
\linkS4class{enriched_results}
}

