% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combine_enrichments.R
\name{combine_text}
\alias{combine_text}
\title{combine text}
\usage{
combine_text(list_characters, names_out, text_id)
}
\arguments{
\item{list_characters}{list containing named character strings}

\item{names_out}{the full list of names to use}
}
\value{
named character vector
}
\description{
Given lists of named character objects, and a character vector of names to be
in the final object, either get the character string from the list that has
the names, or check that the character string is the same across all of the
lists.
}

