% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combine_enrichments.R
\name{combined_coefficient}
\alias{combined_coefficient}
\title{combined coefficient}
\usage{
combined_coefficient(n1, n2)
}
\arguments{
\item{n1}{group 1}

\item{n2}{group 2}
}
\value{
double
}
\description{
takes an average of the \code{overlap} and \code{jaccard} coefficients
}

