% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{combined_enrichment-class}
\alias{combined_enrichment}
\alias{combined_enrichment-class}
\title{combined enrichments}
\description{
The \code{combined_enrichment} class holds the results of combining several
\linkS4class{enriched_result}s together, which includes the original
\linkS4class{enriched_result}s, as well as the \code{annotation_graph}
and combined \linkS4class{annotation} objects.
}
\section{Slots}{

\describe{
\item{\code{enriched}}{list of enriched objects}

\item{\code{enriched_type}}{character describing the enrichment annotation}

\item{\code{annotation}}{\linkS4class{annotation} where the annotation_features
have been combined across the \linkS4class{enriched_results}}

\item{\code{graph}}{the annotation graph, where nodes are annotations, and edges are
weighted by similarity between annotation_features.}
}}

