\name{cwReload-methods}
\docType{methods}
\alias{cwReload}
\alias{cwReload-methods}
\alias{cwReload,numeric,character,ccOptions-method}
\alias{cwReload,numeric,ccOptions-method}
\title{  Methods for Function \code{cwReload} in Package \pkg{categoryCompare} }
\description{
  Methods for function \code{cwReload} in package \pkg{categoryCompare}
}
\section{Methods}{
\describe{

\item{\code{signature(oldCW = "numeric", windowName = "character", ccOpts = "ccOptions")}}{
%%  describe this method here

This method is now deprecated as the \code{cwObj} doesn't store anything, but is
merely the network SUID pointing to the network in Cytoscape. See \code{RCy3::getNetworkSuid}.

}
}}
\keyword{methods}
\keyword{  other possible keyword(s)  }
