% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all_classes.R, R/class_constructors.R
\name{enriched_result}
\alias{enriched_result}
\title{the enriched results class}
\usage{
enriched_result(features, universe, annotation, statistics)

enriched_result(features, universe, annotation, statistics)
}
\arguments{
\item{features}{the features that were differentially expressed (see details)}

\item{universe}{all of the features that were measured}

\item{annotation}{an \linkS4Class{annotation} object}

\item{statistics}{a \linkS4Class{statistical_results} object}
}
\value{
enriched_result
}
\description{
the enriched results class

given all the slots for an \linkS4Class{enriched_result}, checks that all
the data is self-consistent, and creates the \code{enriched_result} object.
}
\section{Slots}{

\describe{
\item{\code{features}}{the "features" of interest, a vector of class ANY}

\item{\code{universe}}{all of the "features" in the background}

\item{\code{annotation}}{list giving the annotation to feature relationship}

\item{\code{statistics}}{a \linkS4class{statistical_results} object}
}}

