% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combine_enrichments.R
\docType{methods}
\name{extract_statistics,statistical_results-method}
\alias{extract_statistics,statistical_results-method}
\title{extract statistics}
\usage{
\S4method{extract_statistics}{statistical_results}(in_results)
}
\arguments{
\item{in_results}{the \linkS4class{statistical_results} object}
}
\value{
data.frame
}
\description{
extract all statistics for a \linkS4class{statistical_results} object. These
can then be combined into a \code{data.frame} that can be returned or used
to annotate the graph of annotations.
}

