% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/significant_results.R
\name{filter_annotation_graph}
\alias{filter_annotation_graph}
\title{filter graph by significant entries}
\usage{
filter_annotation_graph(in_graph, comb_enrich)
}
\arguments{
\item{in_graph}{the \linkS4Class{cc_graph} previously generated}

\item{comb_enrich}{the \linkS4Class{combined_enrichment} that you want to use to filter with}
}
\value{
cc_graph
}
\description{
If a graph has already been generated, it may be faster to filter a previously
generated one than generate a new one from significant data.
}

