% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combine_enrichments.R
\docType{methods}
\name{generate_annotation_graph,combined_enrichment-method}
\alias{generate_annotation_graph,combined_enrichment-method}
\title{generate the annotation graph}
\usage{
\S4method{generate_annotation_graph}{combined_enrichment}(comb_enrichment,
  annotation_similarity = "combined", low_cut = 5, hi_cut = 500)
}
\arguments{
\item{comb_enrichment}{the combined_enrichment object}

\item{annotation_similarity}{which similarity measure to use}

\item{low_cut}{keep only those annotations in the graph with at least this many annotated features}

\item{hi_cut}{keep only those annotations with less than this many annotated features}
}
\value{
\linkS4class{cc_graph}
}
\description{
given a \linkS4class{combined_enrichment}, generate the annotation similarity graph
}

