% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all_generics.R
\name{generate_annotation_graph}
\alias{generate_annotation_graph}
\title{generate annotation graph}
\usage{
generate_annotation_graph(comb_enrichment, annotation_similarity = "combined",
  low_cut = 5, hi_cut = 500)
}
\description{
generate annotation graph
}

