% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combine_enrichments.R
\name{generate_annotation_similarity_graph}
\alias{generate_annotation_similarity_graph}
\title{annotation similarity graph}
\usage{
generate_annotation_similarity_graph(annotation_features,
  similarity_type = "combined")
}
\arguments{
\item{annotation_features}{list where each entry is a set of features to that annotation}

\item{similarity_type}{which type of overlap coefficient to report}
}
\value{
cc_graph
}
\description{
given an annotation-feature list, generate a similarity graph between all of
the annotations
}

