% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graph_visualization.R
\name{generate_colors}
\alias{generate_colors}
\title{generate colors}
\usage{
generate_colors(n_color)
}
\arguments{
\item{n_color}{}
}
\description{
given a bunch of items, generate a set of colors for either single node colorings
or pie-chart annotations. Colors are generated using the \emph{hcl} colorspace,
and for \code{n_color >= 5}, the colors are re-ordered in an attempt to create
the largest contrasts between colors, as they result from being picked on a
circle in \emph{hcl} space.
}

