% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hypergeometric_enrichment.R
\name{generate_link}
\alias{generate_link}
\title{generate link text}
\usage{
generate_link(links)
}
\arguments{
\item{links}{the vector of links}
}
\value{
character
}
\description{
given a named vector of links, generate an actual html link formatted for output in html documents
}

