% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graph_visualization.R
\name{generate_piecharts}
\alias{generate_piecharts}
\title{create piecharts for visualization}
\usage{
generate_piecharts(grp_matrix, use_color)
}
\arguments{
\item{grp_matrix}{the group matrix}

\item{use_color}{the colors for each experiment}
}
\value{
list of png files that are pie graphs
}
\description{
given a group matrix and the colors for each experiment, generate the pie graphs
that will be used as glyphs in Cytoscape
}
\details{
this should \emph{not be exported in the final version}
}

