% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combine_enrichments.R
\docType{methods}
\name{generate_table,combined_enrichment-method}
\alias{generate_table,combined_enrichment-method}
\title{generate table}
\usage{
\S4method{generate_table}{combined_enrichment}(comb_enrichment,
  link_type = "explicit")
}
\arguments{
\item{comb_enrichment}{the combined_enrichment object}

\item{link_type}{should their be an "explicit" link (see details)}
}
\value{
data.frame
}
\description{
given a \linkS4class{combined_enrichment} object, get out the data.frame
either for investigation or to add data to the \linkS4class{annotation_graph}.
}
\details{
the \code{link_type} controls whether to create an "explicit" link
  that is actually a column in the data.frame, or create an "implicit" html link
  that is part of the \code{@name} column in the returned data.frame. Useful
  if you are embedding the data.frame in an html report.
}

