% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/significant_results.R
\docType{methods}
\name{get_significant_annotations,combined_enrichment-method}
\alias{get_significant_annotations,combined_enrichment-method}
\title{get significant annotations}
\usage{
\S4method{get_significant_annotations}{combined_enrichment}(in_results, ...)
}
\arguments{
\item{in_results}{a \linkS4class{combined_enrichment} object}

\item{...}{conditional expressions}
}
\value{
\linkS4class{combined_enrichment} object
}
\description{
In the case where we have a \linkS4class{combined_enrichment} and we want
to get all of the significant annotations from each of them, and put them
together so we can start doing real meta-analysis.
}
\details{
Note that this function returns the original \linkS4class{combined_enrichment} object with a modified
\linkS4class{combined_statistics} slot where the significant annotations have been added in.
}

