% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all_generics.R
\name{get_significant_annotations}
\alias{get_significant_annotations}
\title{get significant annotations}
\usage{
get_significant_annotations(in_results, ...)
}
\description{
get significant annotations
}

