\name{graphType-methods}
\docType{methods}
\alias{graphType-methods}
\alias{graphType}
\alias{graphType<-}
\title{ graphType }
\description{
Gets and sets the \code{graphType} for a couple of different \code{ccEnrichResults} objects
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ccEnrichResult")}}{
%%  describe this method here
}

\item{\code{signature(object = "GENccEnrichResult")}}{
%%  describe this method here
}
}}
\seealso{
\code{\link{ccEnrichResult}} \code{\linkS4class{GENccEnrichResult}}
}
\keyword{methods}
\keyword{  other possible keyword(s)  }
