% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hypergeometric_enrichment.R
\name{hypergeometric_basic}
\alias{hypergeometric_basic}
\title{do hypergeometric test}
\usage{
hypergeometric_basic(num_white, num_black, num_drawn, num_white_drawn,
  direction = "over")
}
\arguments{
\item{num_white}{number of white balls in urn}

\item{num_black}{number of black balls in urn}

\item{num_drawn}{number of balls taken from urn}

\item{num_white_drawn}{number of white balls taken from urn}

\item{direction}{which direction is the test}
}
\value{
list
}
\description{
does a hypergeometric enrichment test
}

