% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hypergeometric_enrichment.R
\name{hypergeometric_feature_enrichment}
\alias{hypergeometric_feature_enrichment}
\title{do hypergeometric enrichment}
\usage{
hypergeometric_feature_enrichment(hypergeom_features, direction = "over")
}
\arguments{
\item{hypergeom_features}{a hypergeometric_features object}

\item{direction}{which direction to do the enrichment (over or under)}
}
\value{
enriched_result
}
\description{
do hypergeometric enrichment
}

