% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combine_enrichments.R
\name{jaccard_coefficient}
\alias{jaccard_coefficient}
\title{jaccard coefficient}
\usage{
jaccard_coefficient(n1, n2)
}
\arguments{
\item{n1}{group 1}

\item{n2}{group 2}
}
\value{
double
}
\description{
calculates similarity of two groups of objects using "jaccard" coefficient,
defined as:
}
\details{
length(intersect(n1, n2)) / min(c(length(n1), length(n2)))
}

