\name{listNames}
\alias{listNames}
\title{
listNames
}
\description{
Extracts the \code{listNames} from \code{\linkS4class{ccGeneList}} or \code{\linkS4class{ccOptions}} objects.
}
\usage{listNames(object)}
\arguments{
	\item{object}{ This will be either a \code{\linkS4class{ccGeneList}} or \code{\linkS4class{ccOptions}} object}
}
\author{
Robert M Flight
}
\seealso{
\code{\linkS4class{ccGeneList}} \code{\linkS4class{ccOptions}}
}

