% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/significant_results.R
\name{multi_query_list}
\alias{multi_query_list}
\title{index a list}
\usage{
multi_query_list(list_to_query, ...)
}
\arguments{
\item{list_to_query}{the list to run the query on}

\item{...}{the expressions that do the queries}
}
\value{
logical "&" of all queries
}
\description{
Provided a list, and a condition, returns the logical indices into the named
part of the list provided. Uses \code{subset} like non-standard evaluation
so that we can define appropriate expressions.
}

