% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{node_assign-class}
\alias{node_assign}
\alias{node_assign-class}
\title{node_assign}
\description{
The \code{node_assign} class holds the unique annotation combinations and the
assignment of the nodes to those combinations for use in visualization.
}
\section{Slots}{

\describe{
\item{\code{groups}}{the unique groups, as a logical matrix}

\item{\code{assignments}}{named character vector providing association with groups}

\item{\code{colors}}{named character vector of hex colors for groups or experiments}

\item{\code{color_type}}{whether doing group or experiment based colors}

\item{\code{pie_locs}}{if doing experiment colors, then pie graphs were generated here}
}}

