% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combine_enrichments.R
\name{overlap_coefficient}
\alias{overlap_coefficient}
\title{overlap coefficient}
\usage{
overlap_coefficient(n1, n2)
}
\arguments{
\item{n1}{group 1 of objects}

\item{n2}{group 2 of objects}
}
\value{
double
}
\description{
calculates the similarity using the "overlap" coefficient, which is
}
\details{
length(intersect(n1, n2)) / length(union(n1, n2))
}

