% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graph_visualization.R
\docType{methods}
\name{remove_edges,cc_graph,numeric-method}
\alias{remove_edges,cc_graph,numeric-method}
\title{remove graph edges}
\usage{
\S4method{remove_edges}{cc_graph,numeric}(edge_obj, cutoff,
  edge_attr = "weight", value_direction = "under")
}
\arguments{
\item{edge_obj}{cc_graph}

\item{cutoff}{the cutoff to use}

\item{edge_attr}{which attribute to use}

\item{value_direction}{remove edges with value under or over}
}
\value{
cc_graph
}
\description{
remove graph edges
}

