% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graph_visualization.R
\docType{methods}
\name{remove_edges,CytoscapeWindowClass,numeric-method}
\alias{remove_edges,CytoscapeWindowClass,numeric-method}
\title{remove edges}
\usage{
\S4method{remove_edges}{CytoscapeWindowClass,numeric}(edge_obj, cutoff,
  edge_attr = "weight", value_direction = "under")
}
\arguments{
\item{edge_obj}{a CytoscapeWindowClass}

\item{cutoff}{what cutoff to use to remove edges}

\item{edge_attr}{what attribute has the values}

\item{value_direction}{remove those edges "under" or "over" the value}
}
\value{
nothing
}
\description{
given a \linkS4Class{CytoscapeWindowClass}, remove edges according to provided
values.
}

