\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,HyperGResultCC-method}
\alias{summary-methods}
\alias{summary,HyperGResultCC-method}
\title{Methods for Function show in Package `categoryCompare'}
\description{
 The \code{show} and \code{summary} methods for \code{\link{HyperGResultCC}} objects generated using \code{\link{hyperGTestCC}}
}
\section{Methods}{
\describe{
\item{\code{show, signature(object = "HyperGResultCC")}}{

}
\item{\code{summary, signature(object = "HyperGResultCC")}}{
}
}
}
\author{
Robert M Flight
}
\examples{
\dontrun{
data(ccData)
show(enrichLists)
summary(enrichLists[[1]][[1]])
}
}
\keyword{methods}

