% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{significant_annotations-class}
\alias{significant_annotations}
\alias{significant_annotations-class}
\title{significant annotations}
\description{
The \code{significant_annotations} class holds which annotations from which
enrichment were both \strong{measured} and \strong{significant}. Each of these
slots is a \emph{logical matrix} with rows named by \emph{annotation_id} and
columns named by the names of the \linkS4class{enriched_result} that was combined.
}
\section{Slots}{

\describe{
\item{\code{significant}}{logical matrix}

\item{\code{measured}}{logical matrix}

\item{\code{sig_calls}}{character representations of calls used to filter the data}
}}

