% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{statistical_results-class}
\alias{statistical_results}
\alias{statistical_results-class}
\title{statistical results class}
\description{
This class holds the part of an enrichment that is the statistical results.
It has two pieces, a \code{list} of \code{statistics} that is a named list
with the actual numerical results of applying the statistics. The other piece
is the \code{annotation_id} vector defining which entry in each vector of the
\code{statistics} is.
}
\section{Slots}{

\describe{
\item{\code{statistic_data}}{list of numerical statistics}

\item{\code{annotation_id}}{vector of ids}
}}

