% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graph_visualization.R
\name{vis_in_cytoscape}
\alias{vis_in_cytoscape}
\title{visualize in cytoscape}
\usage{
vis_in_cytoscape(in_graph, in_assign, description = "", ...)
}
\arguments{
\item{in_graph}{the cc_graph to visualize}

\item{in_assign}{the node_assign generated}

\item{description}{something descriptive about the vis (useful when lots of different visualizations)}

\item{...}{other parameters for \code{CytoscapeWindow}}
}
\value{
something
}
\description{
given a graph, and the node assignments, visualize the graph in cytoscape
for manipulation
}

