\name{cwReload-methods}
\docType{methods}
\alias{cwReload}
\alias{cwReload-methods}
\alias{cwReload,CytoscapeWindowClass,character,ccOptions-method}
\title{  Methods for Function \code{cwReload} in Package \pkg{categoryCompare} }
\description{
  Methods for function \code{cwReload} in package \pkg{categoryCompare} 
}
\section{Methods}{
\describe{

\item{\code{signature(oldCW = "CytoscapeWindowClass", windowName = "character", ccOpts = "ccOptions")}}{
%%  describe this method here
}
}}
\keyword{methods}
\keyword{  other possible keyword(s)  }
